sergeant_service/requirements.txt
numpy
langchain
langchain-community
langchain-openai
langchain_postgres
openai
requests
uvicorn
pydantic
starlette
fastapi
aiohttp
pyyaml
cerberus
anthropic
sentry-sdk[fastapi]
pgvector
sqlalchemy[asyncio]
psycopg2-binary
asyncpg